Added logout support

+++ cinema/src/main/java/controller/CinemaControllerServlet.java
@@ -73,7 +73,7 @@ public class CinemaControllerServlet extends HttpServlet {
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		
 		String command = request.getParameter("command");
-		String username = (String)request.getServletContext().getAttribute("username");
+		String username = (String)getServletContext().getAttribute("username");
 		
 		switch(command) {
 			case "LOGIN": handleLoginRequest(request, response);
@@ -100,13 +100,39 @@ public class CinemaControllerServlet extends HttpServlet {
 								break;
 			case "MANAGER-BOOKING":handleManagerBookingRequest(request, response);
 									break;
+			case "LOGOUT": handleLogoutRequest(request, response);
+							break;
 		}
 	}
 
-	private void handleManagerBookingRequest(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+	private void handleLogoutRequest(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		String username = (String) getServletContext().getAttribute("username");
 		RequestDispatcher requestDispatcher = null;
 		
-		String username = (String) (request.getServletContext().getAttribute("username"));
+		if(username == null) {
+			requestDispatcher = request.getRequestDispatcher("not-authorized.jsp");
+			requestDispatcher.forward(request, response);
+			return;
+		}
+		
+		getServletContext().setAttribute("username", null);
+		
+		requestDispatcher = request.getRequestDispatcher("login-form.html");
+		requestDispatcher.forward(request, response);
+	}
+
+	private void handleManagerBookingRequest(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		String username = (String) getServletContext().getAttribute("username");
+		String role = userService.getRole(username);
+		RequestDispatcher requestDispatcher = null;
+		
+		if(!role.equals("manager")) {
+			requestDispatcher = request.getRequestDispatcher("not-authorized.jsp");
+			requestDispatcher.forward(request, response);
+			return;
+		}
+		
+		
 		
 		String cinema = userService.getCinema(username);
 		
@@ -122,11 +148,20 @@ public class CinemaControllerServlet extends HttpServlet {
 	}
 
 	private void handleLoginRequest(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		String username = (String) getServletContext().getAttribute("username");
 		RequestDispatcher requestDispatcher = null;
-		String username = request.getParameter("username");
+		System.out.println(username);
+		
+		
+		if(username != null) {
+			requestDispatcher = request.getRequestDispatcher("not-authorized.jsp");
+			requestDispatcher.forward(request, response);
+			return;
+		}
+		username = request.getParameter("username");
 		if(userService.areValidCredentials(request.getParameter("username"), request.getParameter("password"))) {
 			
-			request.getServletContext().setAttribute("username", request.getParameter("username"));
+			getServletContext().setAttribute("username", request.getParameter("username"));
 			
 			switch(userService.getRole(username)) {
 				case "admin": request.setAttribute("manager_list", requestService.getRequests());
@@ -150,9 +185,16 @@ public class CinemaControllerServlet extends HttpServlet {
 	
 	
 	private void handleRegisterRequest(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		String username = (String) getServletContext().getAttribute("username");
 		RequestDispatcher requestDispatcher = null;
 		
-		String username = request.getParameter("username");
+		if(username != null) {
+			requestDispatcher = request.getRequestDispatcher("not-authorized.jsp");
+			requestDispatcher.forward(request, response);
+			return;
+		}
+		
+		username = request.getParameter("username");
 		String password = request.getParameter("password");
 		String name = request.getParameter("name");
 		String phone = request.getParameter("phone");
@@ -174,9 +216,17 @@ public class CinemaControllerServlet extends HttpServlet {
 	}
 	
 	private void handleAcceptRequest(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		String username = (String) getServletContext().getAttribute("username");
+		String role = userService.getRole(username);
 		RequestDispatcher requestDispatcher = null;
 		
-		String username = request.getParameter("username");
+		if(!role.equals("admin")) {
+			requestDispatcher = request.getRequestDispatcher("not-authorized.jsp");
+			requestDispatcher.forward(request, response);
+			return;
+		}
+		
+		username = request.getParameter("username");
 		
 		requestService.approveRequest(username);
 		
@@ -187,9 +237,17 @@ public class CinemaControllerServlet extends HttpServlet {
 	}
 	
 	private void handleDenyRequest(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		String username = (String) getServletContext().getAttribute("username");
+		String role = userService.getRole(username);
 		RequestDispatcher requestDispatcher = null;
 		
-		String username = request.getParameter("username");
+		if(!role.equals("admin")) {
+			requestDispatcher = request.getRequestDispatcher("not-authorized.jsp");
+			requestDispatcher.forward(request, response);
+			return;
+		}
+		
+		username = request.getParameter("username");
 		
 		requestService.denyRequest(username);
 		
@@ -200,7 +258,15 @@ public class CinemaControllerServlet extends HttpServlet {
 	}
 	
 	private void handleAddMovieRequest(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		String username = (String) getServletContext().getAttribute("username");
+		String role = userService.getRole(username);
 		RequestDispatcher requestDispatcher = null;
+		
+		if(!role.equals("manager")) {
+			requestDispatcher = request.getRequestDispatcher("not-authorized.jsp");
+			requestDispatcher.forward(request, response);
+			return;
+		}
 		
 		String title = request.getParameter("title");
 		int freeSeats = Integer.parseInt(request.getParameter("freeSeats"));
@@ -208,7 +274,7 @@ public class CinemaControllerServlet extends HttpServlet {
 		int endHour = Integer.parseInt(request.getParameter("endHour"));
 		double price = Double.parseDouble(request.getParameter("price"));
 		
-		String username = (String) (request.getServletContext().getAttribute("username"));
+		username = (String) (getServletContext().getAttribute("username"));
 		
 		String cinema = userService.getCinema(username);
 
@@ -221,7 +287,15 @@ public class CinemaControllerServlet extends HttpServlet {
 	}
 	
 	private void handleAddMovieIdRequest(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		String username = (String) getServletContext().getAttribute("username");
+		String role = userService.getRole(username);
 		RequestDispatcher requestDispatcher = null;
+		
+		if(!role.equals("manager")) {
+			requestDispatcher = request.getRequestDispatcher("not-authorized.jsp");
+			requestDispatcher.forward(request, response);
+			return;
+		}
 		
 		String title = request.getParameter("title");
 		int freeSeats = Integer.parseInt(request.getParameter("freeSeats"));
@@ -230,7 +304,7 @@ public class CinemaControllerServlet extends HttpServlet {
 		double price = Double.parseDouble(request.getParameter("price"));
 		int id = Integer.parseInt(request.getParameter("id"));
 		
-		String username = (String) (request.getServletContext().getAttribute("username"));
+		username = (String) (getServletContext().getAttribute("username"));
 		
 		String cinema = userService.getCinema(username);
 
@@ -243,7 +317,15 @@ public class CinemaControllerServlet extends HttpServlet {
 	}
 	
 	private void handleUpdateMovieRequest(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		String username = (String) getServletContext().getAttribute("username");
+		String role = userService.getRole(username);
 		RequestDispatcher requestDispatcher = null;
+		
+		if(!role.equals("manager")) {
+			requestDispatcher = request.getRequestDispatcher("not-authorized.jsp");
+			requestDispatcher.forward(request, response);
+			return;
+		}
 		String id = request.getParameter("id");
 		
 		Movie movie = movieService.findMovie(Integer.parseInt(id));
@@ -258,12 +340,20 @@ public class CinemaControllerServlet extends HttpServlet {
 	}
 	
 	private void handleDeleteMovieRequest(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		String username = (String) getServletContext().getAttribute("username");
+		String role = userService.getRole(username);
 		RequestDispatcher requestDispatcher = null;
+		
+		if(!role.equals("manager")) {
+			requestDispatcher = request.getRequestDispatcher("not-authorized.jsp");
+			requestDispatcher.forward(request, response);
+			return;
+		}
 		String id = request.getParameter("id");
 		
 		movieService.deleteMovie(Integer.parseInt(id), false);
 		
-		String username = (String) (request.getServletContext().getAttribute("username"));
+		username = (String) (getServletContext().getAttribute("username"));
 		
 		String cinema = userService.getCinema(username);
 
@@ -275,7 +365,15 @@ public class CinemaControllerServlet extends HttpServlet {
 	}
 
 	private void handleSeeMoviesRequest(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		String username = (String) getServletContext().getAttribute("username");
+		String role = userService.getRole(username);
 		RequestDispatcher requestDispatcher = null;
+		
+		if(!role.equals("regular")) {
+			requestDispatcher = request.getRequestDispatcher("not-authorized.jsp");
+			requestDispatcher.forward(request, response);
+			return;
+		}
 		String cinema = request.getParameter("cinema");
 		
 		request.setAttribute("movie_list", movieService.getMoviesFromCinema(cinema));
@@ -285,7 +383,15 @@ public class CinemaControllerServlet extends HttpServlet {
 	}
 	
 	private void handleBookMovieRequest(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		String username = (String) getServletContext().getAttribute("username");
+		String role = userService.getRole(username);
 		RequestDispatcher requestDispatcher = null;
+		
+		if(!role.equals("regular")) {
+			requestDispatcher = request.getRequestDispatcher("not-authorized.jsp");
+			requestDispatcher.forward(request, response);
+			return;
+		}
 		int id = Integer.parseInt(request.getParameter("id"));
 		
 		request.setAttribute("id", id);
@@ -295,12 +401,20 @@ public class CinemaControllerServlet extends HttpServlet {
 	}
 	
 	private void handleBookRequest(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+		String username = (String) getServletContext().getAttribute("username");
+		String role = userService.getRole(username);
 		RequestDispatcher requestDispatcher = null;
+		
+		if(!role.equals("regular")) {
+			requestDispatcher = request.getRequestDispatcher("not-authorized.jsp");
+			requestDispatcher.forward(request, response);
+			return;
+		}
 		int id = Integer.parseInt(request.getParameter("id"));
 		int places = Integer.parseInt(request.getParameter("selectedSeats"));
 		
 		Movie movie = movieService.findMovie(id);
-		String username = (String) (request.getServletContext().getAttribute("username"));
+		username = (String) (getServletContext().getAttribute("username"));
 		
 		try {
 			bookingService.addBooking(new Booking(username, id, places));
@@ -316,6 +430,8 @@ public class CinemaControllerServlet extends HttpServlet {
 		
 		requestDispatcher.forward(request, response);
 	}
+	
+	
 	
 	/**
 	 * @see HttpServlet#doPost(HttpServletRequest request, HttpServletResponse response)
